Load and attach kfunc/kretfunc programs
Adding support to specify kfunc/kretfunc attach points.

The 'kfunc' attach point is a bpftrace 'code name' for
function trampolines, which is a new way to attach BPF
program to kernel functions.

The trampolines are closely binded with BTF kernel data,
and that allows them to know the name and type of arguments.

We will be able to use function arguments by arg->NAME,
where NAME is the function argument name, like:

  # bpftrace -e 'kretfunc:fget { printf("fd %d\n", args->fd); }'
  Attaching 1 probe...
  fd 3
  ...

We are also able to access return value via args->ret
(together with kfunc function arguments), like:

  # bpftrace -e 'kretfunc:fget { printf("fd %d name %s\n", args->fd, str(args->ret->f_path.dentry->d_name.name));  }'
  Attaching 1 probe...
  fd 9 name _journal.cpython-37m-x86_64-linux-gnu.so
  ...

As I already mentioned and presented above the kretfunc has
access (same way as kfunc) to function arguments as well.
This will speed up some scripts, that need function arguments
at the return probe.

The kfunc and kretfunc is also a starting point for using BTF
powered argument passing to kernel helpers, which would allow
us to do fast filepath helper, that operates on 'struct file'
pointer directly. I have the code ready both for kernel and
bpftrace and it will follow once this change in.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>